+++ src/Model/PaazlManagement.php
@@ -157,6 +157,9 @@ class PaazlManagement implements \Paazl\Shipping\Api\PaazlManagementInterface
      */
     public function processOrderCommitRequest($order)
     {
+        /**
+         * @var $order \Magento\Sales\Model\Order
+         */
         $shippingMethod = $order->getShippingMethod(true);
         $shippingAddress = $order->getShippingAddress();
         /**
@@ -190,11 +193,12 @@ class PaazlManagement implements \Paazl\Shipping\Api\PaazlManagementInterface
                 'shippingMethod' => [
                     'type' => 'delivery',
                     'identifier' => null,
-                    'option' => $shippingMethod->getMethod(),
+                    'option' => $rate['method'],
                     'orderWeight' => $this->getConvertedWeight($order->getWeight()),
                     'description' => 'Delivery', //@todo Find out what description is expected
                     'assuredAmount' => $assuredAmount,
-                    'assuredAmountCurrency' => 'EUR'
+                    'assuredAmountCurrency' => 'EUR',
+                    'customsValue' => $order->getSubtotal(),
                 ],
                 'shippingAddress' => [
                     'customerName' => $shippingAddress->getName(),