+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Bugfix issue radio buttons become disabled after page refresh. See: https://github.com/magento/magento2/issues/7497
+- Set correct shipping method coming from Paazl
+- Add customsValue for Fedex orders
 
 ## [1.0.5] - 2017-04-26
 ### Changed